Full English language support
Full English language support has been added using resource files. They are tables of keys with values. This allows you to translate the program into any other language in the future!
The location of the title on the title frame has also been slightly changed and font sizes have been changed depending on the size of the window.